+++ CHANGELOG.md
@@ -10,6 +10,13 @@ Welcome to the Github! If you are looking for the most recent stable version, th
 
 ## (V.1.6.6 Changes) (1.16.3 Minecraft)
 
+##### Misc:
+
+- Added check to automatically blacklist any dimension using the
+  FlatChunkGenerator (Superflat worldtype) because that chunk generator
+  will spawn all RS's structures at once and make /locate on certain 
+  RS structure cause the server to hang forever.
+
 ##### Backend:
 
 - Cleaned up codebase a bit.